Deleted unused code
Deleted unused coments and components. Renamed Drawer to Sidebar. Touched up some styling code,
