--- client/src/components/Profile/Courses.js
@@ -1,13 +1,0 @@
-import React from 'react';
-import { Grid } from '@material-ui/core';
-
-const Courses = () => {
-    return (
-        // Courses container
-        <Grid container>
-            <h1>This is courses</h1>
-        </Grid>
-    );
-};
-
-export default Courses;
