+++ client/src/components/Profile/Sidebar.js
@@ -1,85 +1,115 @@
 import React from 'react';
 import { Grid, Typography } from '@material-ui/core';
+import { Drawer as MUIDrawer } from '@material-ui/core';
 import { NavLink } from 'react-router-dom';
 import { makeStyles } from '@material-ui/styles';
-import Avatar from '@material-ui/core/Avatar';
+import profilePic from '../../static/images/profilePicSample.png';
 
-const useStyles = makeStyles((theme) => ({
-    testStyles: {
-        padding: '1em 0',
+const useStyles = makeStyles(theme => ({
+    linkStyles: {
+        textDecoration: 'none',
+        color: 'black',
+        padding: '1em 0 0 2em',
+        fontSize: 18,
+        opacity: '50%',
     },
-    navStyles: {
-        padding: '1em 0',
+    profilePic: {
+        borderRadius: '50%',
+        width: '130px',
+        marginTop: '4em',
     },
-    large: {
-        width: theme.spacing(15),
-        height: theme.spacing(15),
+    drawer: {
+        width: 400,
+        background: '#F9F9FC',
     },
-    userName: {
+    linkContainer: {
+        paddingTop: '2em',
+    },
+    profileName: {
         paddingTop: '1em',
-        fontWeight: 'bold',
+        fontSize: 22,
     },
 }));
 
-const Sidebar = () => {
+const Drawer = props => {
     const classes = useStyles();
 
     return (
-        //  Container for Profile Summary
-        <Grid item container={12} direction='column'>
-            {/* Container to hold user info (picture/name) */}
-            <Grid item container direction='column' sm={6}>
-                <Grid item style={{ paddingTop: '2em 0 5em 0' }} alignSelf='center'>
-                    <Avatar
-                        alt='Ashley Sanford'
-                        src='/static/images/avatar/3.jpg'
-                        className={classes.large}
+        <MUIDrawer variant="permanent" anchor="left" className={classes.drawer}>
+            {/* Drawer container */}
+            <Grid
+                container
+                direction="column"
+                alignItems="center"
+                style={{ width: '400px' }}
+                sm={12}
+                className={classes.drawer}
+            >
+                {/* Profile briefing container */}
+                <Grid container item direction="column" alignItems="center" sm={4}>
+                    <img
+                        src={profilePic}
+                        alt="Ashly Sanford"
+                        className={classes.profilePic}
                     />
-                    <Typography className={classes.userName}>Ashley Sanford</Typography>
+
+                    <Typography className={classes.profileName} align="center">
+                        Ashley Sanford
+                    </Typography>
+                </Grid>
+
+                {/* Profile Links container */}
+                <Grid container item direction="column" className={classes.linkContainer}>
+                    <NavLink
+                        to="/profile"
+                        exact
+                        activeStyle={{
+                            opacity: '100%',
+                        }}
+                        className={classes.linkStyles}
+                    >
+                        User Info
+                    </NavLink>
+                    <NavLink
+                        to="profile/courses"
+                        exact
+                        activeStyle={{
+                            opacity: '100%',
+                        }}
+                        className={classes.linkStyles}
+                    >
+                        My Courses
+                    </NavLink>
+                    <NavLink
+                        to="/profile/settings"
+                        exact
+                        activeStyle={{
+                            opacity: '100%',
+                        }}
+                        className={classes.linkStyles}
+                    >
+                        Settings
+                    </NavLink>
+                    <NavLink
+                        to="/profile/notifications"
+                        exact
+                        activeStyle={{
+                            opacity: '100%',
+                        }}
+                        className={classes.linkStyles}
+                    >
+                        Notifications
+                    </NavLink>
+                </Grid>
+                <Grid container direction="column" item justify="flex-end" sm={4}>
+                    <Grid item>
+                        <Typography align="left">Logout</Typography>
+                    </Grid>
                 </Grid>
             </Grid>
-            {/* Container to hold router links */}
-            <Grid item container direction='column' justify='center'>
-                <NavLink
-                    to='/profile/'
-                    exact
-                    activeStyle={{
-                        fontWeight: 'bold',
-                    }}
-                    className={classes.linkStyles}
-                >
-                    User Info
-                </NavLink>
-                <NavLink
-                    to='/profile/courses'
-                    activeStyle={{
-                        fontWeight: 'bold',
-                    }}
-                    className={classes.linkStyles}
-                >
-                    My Courses
-                </NavLink>
-                <NavLink
-                    to='/profile/settings'
-                    activeStyle={{
-                        fontWeight: 'bold',
-                    }}
-                    className={classes.linkStyles}
-                >
-                    Settings
-                </NavLink>
-                <NavLink
-                    to='/profile/notifications'
-                    activeStyle={{
-                        fontWeight: 'bold',
-                    }}
-                    className={classes.linkStyles}
-                >
-                    Notifications
-                </NavLink>
-            </Grid>
-        </Grid>
+        </MUIDrawer>
+        //  Container for Profile Summary
     );
 };
 
-export default Sidebar;
+export default Drawer;
